requirements.txt
torch
lightning
sentence-transformers
lightgbm
faiss-cpu
qdrant-client
fastapi
uvicorn
mlflow
crewai
dspy-ai
bitsandbytes
transformers
docker
pandas
tqdm
pyarrow
